Ploting the most optimised value for P0
Sooo…I messed up last repo, it wasn't returning more accurate values of P0, it was returning the smallest value for chi**2, this modification makes it return the corresponding most accurate value for P0, I realised it kept returning the lowest possible value for P0 that was in the searching range, so I lower the range that the program searched in, and it returned about 0.75e+08, so I made the program search for a value more precise around that area. I also made the program re-create the predicted velarray, and plot it.